requirements.txt
aiogram>=3.13,<4
httpx>=0.27,<0.29
python-dotenv>=1.0,<2
ebooklib>=0.18,<0.20
lxml>=5.0,<6
